# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestPayload": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "league_type_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League Type Code"
          },
          "player_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Id"
          },
          "season_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Year"
          },
          "source_row_id": {
            "title": "Source Row Id",
            "type": "string"
          },
          "source_table": {
            "title": "Source Table",
            "type": "string"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Uri"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content",
          "source_table",
          "source_row_id"
        ],
        "title": "IngestPayload",
        "type": "object"
      },
...
      "RunIngestPayload": {
        "properties": {
          "mode": {
            "default": "INCREMENTAL",
            "title": "Mode",
            "type": "string"
          },
          "season_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Year"
          },
          "since": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          },
          "tables": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tables"
          },
          "trigger_source": {
            "default": "MANUAL_API",
            "title": "Trigger Source",
            "type": "string"
          }
        },
        "title": "RunIngestPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "InternalApiKey": {
        "description": "AI 내부 호출용 키. Authorization Bearer 토큰을 사용할 수 있습니다.",
        "in": "header",
        "name": "X-Internal-Api-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "KBO AI Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ai/ingest/": {
      "post": {
        "operationId": "ingest_document_ai_ingest__post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Internal-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "InternalApiKey": []
          }
        ],
        "summary": "Ingest Document",
        "tags": [
          "ingest"
        ]
      }
    },
    "/ai/ingest/run": {
      "post": {
        "description": "Persist or deduplicate a durable internal-DB ingestion run.",
        "operationId": "run_ingestion_job_ai_ingest_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Internal-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunIngestPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "InternalApiKey": []
          }
        ],
        "summary": "Run Ingestion Job",
        "tags": [
          "ingest"
        ]
      }
    },
    "/ai/ingest/runs/{run_id}": {
      "get": {
        "description": "Return only sanitized durable run status fields.",
        "operationId": "get_ingestion_run_ai_ingest_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Internal-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "InternalApiKey": []
          }
        ],
        "summary": "Get Ingestion Run",
        "tags": [
          "ingest"
        ]
      }
    },
...
  }
}